Working Week 6 iteration.
